+++ test/functional/Deferred.hs
@@ -12,7 +12,7 @@ import Language.Haskell.LSP.Test
 import Language.Haskell.LSP.Types
 import Language.Haskell.LSP.Types.Lens hiding (id, message)
 import qualified Language.Haskell.LSP.Types.Lens as LSP
-import Test.HIE.Util
+import Test.Hls.Util
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Hspec.Expectations
