+++ test/functional/Diagnostic.hs
@@ -13,7 +13,7 @@ import           Ide.Plugin.Config
 import           Language.Haskell.LSP.Test hiding (message)
 import           Language.Haskell.LSP.Types
 import qualified Language.Haskell.LSP.Types.Lens as LSP
-import           Test.HIE.Util
+import           Test.Hls.Util
 import           Test.Tasty
 import           Test.Tasty.HUnit
 import           Test.Hspec.Expectations
