+++ test/functional/FunctionalBadProject.hs
@@ -8,7 +8,7 @@ module FunctionalBadProject (tests) where
 -- import           Language.Haskell.LSP.Test hiding (message)
 -- import           Language.Haskell.LSP.Types as LSP
 -- import           Language.Haskell.LSP.Types.Lens as LSP hiding (contents, error )
--- import Test.HIE.Util
+-- import Test.Hls.Util
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Hspec.Expectations
