+++ test/functional/FunctionalCodeAction.hs
@@ -20,7 +20,7 @@ import           Language.Haskell.LSP.Test as Test
 import           Language.Haskell.LSP.Types
 import qualified Language.Haskell.LSP.Types.Lens as L
 import qualified Language.Haskell.LSP.Types.Capabilities as C
-import           Test.HIE.Util
+import           Test.Hls.Util
 import           Test.Tasty
 import           Test.Tasty.HUnit
 import           Test.Hspec.Expectations
