+++ test/functional/FunctionalLiquid.hs
@@ -11,7 +11,7 @@ import           Language.Haskell.LSP.Test hiding (message)
 import           Language.Haskell.LSP.Types as LSP
 import           Language.Haskell.LSP.Types.Lens as LSP hiding (contents, error )
 import           Ide.Plugin.Config
-import           Test.HIE.Util
+import           Test.Hls.Util
 import           Test.Tasty
 import           Test.Tasty.HUnit
 import           Test.Hspec.Expectations
