+++ test/functional/Highlight.hs
@@ -5,7 +5,7 @@ import Control.Applicative.Combinators
 import Control.Monad.IO.Class
 import Language.Haskell.LSP.Test
 import Language.Haskell.LSP.Types
-import Test.HIE.Util
+import Test.Hls.Util
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Hspec.Expectations
