+++ test/functional/Progress.hs
@@ -12,7 +12,7 @@ import Language.Haskell.LSP.Messages -- TODO: Move this into haskell-lsp-types
 import Language.Haskell.LSP.Types
 import qualified Language.Haskell.LSP.Types.Lens as L
 import Language.Haskell.LSP.Types.Capabilities
-import Test.HIE.Util
+import Test.Hls.Util
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Hspec.Expectations
