+++ test/functional/Rename.hs
@@ -4,7 +4,7 @@ module Rename (tests) where
 -- import Control.Monad.IO.Class
 -- import Language.Haskell.LSP.Test
 -- import Language.Haskell.LSP.Types
--- import Test.HIE.Util
+-- import Test.Hls.Util
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Hspec.Expectations
