+++ test/functional/Symbol.hs
@@ -5,7 +5,7 @@ import Control.Monad.IO.Class
 import Language.Haskell.LSP.Test as Test
 import Language.Haskell.LSP.Types
 import Language.Haskell.LSP.Types.Capabilities
-import Test.HIE.Util
+import Test.Hls.Util
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Hspec.Expectations
